Day 24 file Input and Output
Modification of the snake game, so that it's able to read and write
the max score achieved. The code used was Angela's instead of mine, so
I could follow what she was doing, instead of getting entangled in the
differences of code.

Then I worked in a mail merge project to hone the I/O skills.