Days 27 and 28 Tkinter
On day 27 tkinter was introduced and a small app was developed to
convert from miles to km.
On day 28 a pomodoro app was developed with tkinter. It was a lot
harder than the previous challenges and apps, and I felt lost at
several points. There's still one thing that I can't quite understand:
a recursive call to a function.