Day 32 smptlib and datetime
This commit has some data files of day 31 (I improved the original
word list), and code for day 32 (smptlib and datetime). The challenge
was to write a script that emails a random message (a letter or a
quote) to a specific person given a set of conditions (birthday or monday).