Day 35 API keys and authentication
Today I wrote a small app to text me if there's rain in the outlook
for the next 12 hours in a given location. The goal of the project was
to connect to two different APIs (that required authentication): one
for the weather and another one with twilio to send SMS. The
conditions of twilio have changed since the lecture was created, so I
used telegram's API (and a bot) to send a text message.