Day 38 Workout Tracker App
I developed today's project with very little help from Angela.
Initially, I wrote all the code in main.py (as Angela did), but then I
moved to OO and wrote classes for nutrition, exercise and sheety.
This could be integrated with telegram in the future. But so far, I've
felt good.