Day 39 Flight Deal Finder app
I worked with an API from kiwi and sheety (to read a set of dream
destinations - prices established by the user) and telegram to notify
me, if any of my conditions was satisfied (that is, if a cheap flight
was found). I went all alone, doing the project in OO and using list
and dictionary comprehension as much as I could. I still struggle with
list and dictionary methods (such as all, any, join) and still have
the R mindset towards a lot of the data processing tasks.