Day 67 blog capstone project
I didn't complete this project. It's been more than a month since I
completed day 66 and I was completely lost with this code. I tried to
follow Angela's code, but it was so buggy it posed more problems than
it solved.

So I decided to keep marching. If I ever need the concepts of this
session, for sure I'll be forced to come back to it.